Remove in-tree taggit serializer
In-tree copy of django-taggit-serializer was introduced in
77111120359a2ef3e4ee6e33e7990b5a62261a29 as a temporary workaround for
problems with running on top of Django 4.0.

django-taggit decided to "vendor" a patched copy of latest
django-taggit-serializer snapshot as part of their PR #738. That was
made available in django-taggit 1.5.0, released in June 2021.

Since django-taggit vendoring clearly predates commit here, it's obvious
we could have used that solution from the beginning. I must have missed
it.